+++ code/src/cljs/sixsq/nuvla/ui/deployment/views.cljs
@@ -152,14 +152,12 @@
 
 
 (defn row-fn
-  [{:keys [id state module parent] :as deployment}]
-  (let [credential-id parent
-        creds-name    (subscribe [::subs/creds-name-map])
-        [primary-url-name
+  [{:keys [id state module credential-name parent] :as deployment}]
+  (let [[primary-url-name
          primary-url-pattern] (-> module :content (get :urls []) first)
-        url           @(subscribe [::subs/deployment-url id primary-url-pattern])
-        select-all?   (subscribe [::subs/select-all?])
-        selected?     (subscribe [::subs/is-selected? id])]
+        url         @(subscribe [::subs/deployment-url id primary-url-pattern])
+        select-all? (subscribe [::subs/select-all?])
+        selected?   (subscribe [::subs/is-selected? id])]
     [ui/TableRow
      (when-not @select-all?
        [ui/TableCell
@@ -180,7 +178,7 @@
      [ui/TableCell (-> deployment :created time/parse-iso8601 time/ago)]
      [ui/TableCell {:style {:overflow      "hidden",
                             :text-overflow "ellipsis",
-                            :max-width     "20ch"}} (get @creds-name credential-id credential-id)]
+                            :max-width     "20ch"}} (or credential-name parent)]
      [ui/TableCell
       (cond
         (utils-general/can-operation? "stop" deployment)
@@ -221,21 +219,20 @@
 
 
 (defn DeploymentCard
-  [{:keys [id state module tags] :as deployment}]
+  [{:keys [id state module tags parent credential-name] :as deployment}]
   (let [tr            (subscribe [::i18n-subs/tr])
-        creds-name    (subscribe [::subs/creds-name-map])
         credential-id (:parent deployment)
         {module-logo-url :logo-url
          module-name     :name
          module-content  :content} module
-        cred-info     (get @creds-name credential-id credential-id)
         [primary-url-name
          primary-url-pattern] (-> module-content (get :urls []) first)
         primary-url   (subscribe [::subs/deployment-url id primary-url-pattern])
         started?      (utils/is-started? state)
         dep-href      (utils/deployment-href id)
         select-all?   (subscribe [::subs/select-all?])
-        is-selected?  (subscribe [::subs/is-selected? id])]
+        is-selected?  (subscribe [::subs/is-selected? id])
+        cred          (or credential-name parent)]
     ^{:key id}
     [uix/Card
      (cond-> {:header        [:span [:p {:style {:overflow      "hidden",
@@ -243,8 +240,8 @@
                                                  :max-width     "20ch"}} module-name]]
               :meta          (str (@tr [:created]) " " (-> deployment :created
                                                            time/parse-iso8601 time/ago))
-              :description   (when-not (str/blank? cred-info)
-                               [:div [ui/Icon {:name "key"}] cred-info])
+              :description   (when cred
+                               [:div [ui/Icon {:name "key"}] cred])
               :tags          tags
               :button        (when (and started? @primary-url)
                                [ui/Button {:color    "green"
@@ -272,9 +269,8 @@
               :state         state
               :loading?      (utils/deployment-in-transition? state)}
 
-             (not @select-all?)
-             (assoc :on-select #(dispatch [::events/select-id id])
-                    :selected? @is-selected?))]))
+             (not @select-all?) (assoc :on-select #(dispatch [::events/select-id id])
+                                       :selected? @is-selected?))]))
 
 
 (defn cards-data-table